requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
pydantic-settings==2.1.0
